restructuring, and fixing references
