+++ common/common-resources.ts
@@ -1,5 +1,5 @@
 import { Snowflake, Client } from 'discord.js';
-import { attach, FallbackType, Config, radios, defaultRadio, MusicData } from './internal.js';
+import { attach, FallbackType, Config, radios, defaultRadio, MusicData } from '../internal.js';
 import { Umzug, SequelizeStorage } from 'umzug';
 import sequelize from 'sequelize';
 const { Sequelize, QueryTypes, DataTypes } = sequelize; //Workaround (CommonJS -> ES modul)
