+++ internal.ts
@@ -1,10 +1,10 @@
-export * from './common-types.js';
-export * from './vc-constants.js';
-export * from './common-resources.js';
-export * from './config-manager.js';
-export * from './util.js';
-export * from './soundcloud-util.js';
-export * from './guild-player.js';
-export * from './actions.js';
-export * from './vc-decorators.js';
-export * from './commands.js';
+export * from './common/common-types.js';
+export * from './common/vc-constants.js';
+export * from './common/common-resources.js';
+export * from './common/config-manager.js';
+export * from './common/util.js';
+export * from './player/soundcloud-util.js';
+export * from './player/guild-player.js';
+export * from './actions/actions.js';
+export * from './actions/vc-decorators.js';
+export * from './actions/commands.js';
